requirements.txt
matplotlib
pandas
numpy
seaborn
python-dotenv
facenet_pytorch
moviepy
pillow
tqdm
transformers
chardet
python-magic
pydub
streamlit
Flask
flask-cors
